% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{app}
\alias{app}
\title{Launch Shiny App}
\usage{
app(...)
}
\arguments{
\item{...}{arguments to pass to shiny::runApp}
}
\value{
Runs the app.
}
\description{
Launch Shiny App
}
\examples{
\dontrun{
app() # to run the app without special options
app(launch.browser = T) # to run the app in a new browser session
}
}
