% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlockmodelingGUI-package.R
\docType{package}
\name{BlockmodelingGUI-package}
\alias{BlockmodelingGUI}
\alias{BlockmodelingGUI-package}
\title{BlockmodelingGUI: GUI for the Generalised Blockmodeling of Valued Networks}
\description{
This app provides some useful tools for Offering an accessible GUI for generalised blockmodeling of single-relation, one-mode networks. The user can execute blockmodeling without having to write a line code by using the app's visual helps. Moreover, there are several ways to visualisations networks and their partitions. Finally, the results can be exported as if they were produced by writing code. The development of this package is financially supported by the Slovenian Research Agency (www.arrs.gov.si) within the research project J5-2557 (Comparison and evaluation of different approaches to blockmodeling dynamic networks by simulations with application to Slovenian co-authorship networks).
}
\author{
\strong{Maintainer}: Fabio Ashtar Telarico \email{telaricof@fdv.uni-lj.si} (\href{https://orcid.org/0000-0002-8740-7078}{ORCID})

Authors:
\itemize{
  \item Aleš Žiberna
}

}
\keyword{internal}
