% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCitationFile.R
\name{GetCitationFile}
\alias{GetCitationFile}
\title{Get citation files}
\usage{
GetCitationFile(filetype = c("ris", "BibText"))
}
\arguments{
\item{filetype}{Type of blibliographic file}
}
\description{
Get citation files
}
