% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HowToCite.R
\name{HowToCite}
\alias{HowToCite}
\title{Show citation info}
\usage{
HowToCite(...)
}
\arguments{
\item{...}{Options to pass to cat()}
}
\value{
Returns information on how to cite the app.
}
\description{
Show citation info
}
\examples{
HowToCite() # to see information about citations
}
